pyproject.toml
[tool.poetry]
name = "CCDCOE"
version = "0.0.5"
description = "Package with general devops code"
readme = "README.md"
repository = "https://github.com/ccdcoe/ccdcoe"
authors = ["CCDCOE <ccdcoe@invalid.com>"]
license = "GNU General Public License v3.0"
classifiers=[
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]


[tool.poetry.urls]
"Code" = "https://github.com/ccdcoe/ccdcoe"


[tool.ccdcoe.group.dependencies]
gitlab_bot = ["common", "deployments", "http_apis", "loggers"]
cli_code = ["cli","common", "deployments", "http_apis", "loggers", "dumpers"]
deployment_code = ["common", "deployments", "http_apis", "loggers"]

[tool.poetry.group.common.dependencies]
python-gitlab = "^5.6.0"
python-dotenv = "^1.0.1"
dataclasses-json = "^0.6.7"

[tool.poetry.group.http_apis.dependencies]
requests = ">=2.31.0"

[tool.poetry.group.loggers.dependencies]
ansicolors = ">=1.1.8"

[tool.poetry.group.dumpers.dependencies]
pyyaml = "^6.0.2"

[tool.poetry.group.cli.dependencies]
tabulate = "^0.9.0"
click = "^8.1.8"

[tool.poetry.group.deployments.dependencies]
tqdm = "^4.67.1"

[tool.poetry.dependencies]
python = "^3.10"
setuptools = ">=68.1.0"

[build-system]
requires = ["setuptools >= 61.0", "poetry == 1.8.4"]
build-backend = "setuptools.build_meta"

[tool.bandit]
exclude_dirs = ["tests"]
tests = []
skips = []
